Fix spell checking function by loading 'cl
When I tried to spell-check a single word with M-$ a few minutes ago,
I was given the Emacs error "Symbol's function definition is void:
remove-if" because Emacs on my newly re-installed system, apparently,
does not happen to load the 'cl Common Lisp module during startup, as
Emacs on my old install must have done.  So, now, I load 'cl manually
so that the remove-if function is present and working again.