Create a dedicated mode for Git commit messages
Previously, I would have Emacs detect a "COMMIT_EDITMSG" filename and
go into "diff" mode, so that it would syntax-highlight the diffs that
I have git attach to each of my commit message buffers for review
while writing up the commit message.  But I have started to miss the
features that I enjoy in my other text buffers, like flyspell
spell-check highlighting and word wrapping.  So with this change I now
have a dedicated git-commit-mode of my own devising, that is basically
diff-mode with the extra features that I need turned on.