Replace recentf with find-file-in-repository
It takes way too long for recentf to learn which files I use, and if my
computer happens to shut down without my closing Emacs, then it loses
all of the files that it learned about during that particular session.
A shortcut to ido-mode opening of any file in the repository is much
better on large code bases like those in which I now work.