Remove git-commit-mode
This function I wrote, to create a derived mode for editing git commits,
has started giving an error about the wrong number of arguments to
"define-derived-mode" when I try to commit.  The error throws me out of
emacsclient and gets printed to the screen with the terminal settings
askance.  The really confusing thing is that, having removed this mode,
the diffs that come up when I run "git commit" are still syntax
highlighted!  Does Emacs now auto-detect in Text Mode when a file looks
like a diff, and applies highlighting?  In any case, diff syntax
highlighting (red "-" lines and green "+" lines) is the only purpose for
which I had created this mode, so I need it no longer, even apart from
the fact that it now throws an error under Ubuntu 15.10.