Make M-p and M-n include digits, too
That way, I can still jump forward and back between symbols that have
digits in the middle (or even just numbers).  Also, while I’m in here,
fix the off-by-one error that was positioning the cursor on the final
letter of the word when moving forward, instead of at its end.