Make k&r the default C style to avoid “gnu” horror
The default style is awful.  It wants to indent lines to the right of
the brace on the preceding line:
```
chap95_rec chap95_jupiter[] = {
                               {-3658015942., -908357166., -300271978}, ...
```
This is a first step towards getting automatic indent detection working,
since the whole exercise is rather pointless if the default indent is
ignored because indentation is chosen based on where a brace falls.