Abandon M-C-a; instead, add interactivity to M-a
Using M-C-a for “let me type a regex to search this project for” was
failing in any mode that instead uses M-C-a for something useless like
“jump to beginning of definition” (C Mode was the most recent offender;
you can see in the code deleted here that JS mode was the previous one).

So instead, let’s give up on that key combination, and instead put a
backdoor into M-a: if it’s called without a symbol at point, then
instead of dying (like it used to), it now prompts for a regex.  Nice!