Stop letting prefix match turn off flex matching
Because the first completion style to return any results ‘wins’ and
stops the search, prefix matches were prevening me from finding things
through flex matching.  This was particularly terrible for M-x, because
sometimes the key word I remember is in the middle of a command name
rather than at the beginning.